merge before you do anything else

+++ HotelReservationSystem/HotelReservationSystem-ejb/src/java/ejb/session/stateless/RoomControllerLocal.java
@@ -21,11 +21,11 @@ public interface RoomControllerLocal {
 
     public List<Room> retrieveAllRooms();
 
-    public Room retrieveRoomById(Integer roomId, Boolean fetchRoomType, Boolean fetchReservation) throws RoomNotFoundException;
+    public Room retrieveRoomById(Long roomId, Boolean fetchRoomType, Boolean fetchReservation) throws RoomNotFoundException;
 
     public void updateRoom(Room room, String roomTypeName, Long reservationLineItemId) throws RoomNotFoundException, RoomTypeNotFoundException, ReservationLineItemNotFoundException;
 
-    public void deleteRoom(Integer roomId) throws RoomNotFoundException, DeleteRoomException;
+    public void deleteRoom(Long roomId) throws RoomNotFoundException, DeleteRoomException;
 
     public List<Room> retrieveRoomByRoomType(Long roomTypeId);
